Add a watchdog task that runs frequently to check for computer returning from sleep. Because timers run at a reduced rate when a task is backgrounded, the interval is set to 1s This makes it a high priority against other interval tasks. In the future I'd like to implement an extension to the dispatcher that runs a collection of tasks at an interval. Reducing this issue.
