+++ src/components/utils.js
@@ -14,5 +14,40 @@ module.exports = {
     }
 
     return url;
+  },
+  /**
+   * Reload when computer wakes up
+   * Tracks time difference, if the difference is greater than 3 minutes. The window reloads
+   * If we know how long Messenger's session takes to timeout, perhaps use that time instead?
+   */
+  watchComputerWake: function(win, winBehaviour) {
+    var time = new Date().getTime();
+    setInterval(function() {
+      var currentTime = new Date().getTime();
+      var deltaTime = currentTime - time;
+      if (deltaTime > 60000 * 3) {  // Time in ms. 3 minutes.
+        // Do the reload
+        winBehaviour.saveWindowState(win);
+        win.reload();
+        /**
+         * Rerun the online check.
+         * We might have changed locations and no longer have internet access
+         */
+        this.checkForOnline(win);
+      }
+	  time = currentTime;
+    }, 1000);  // When the window is backgrounded, all functions run once per second.
+  },
+  /**
+   * Reloads the app once every 10 seconds until the browser is in online mode.
+   */
+  checkForOnline: function(win) {
+    var reloadIntervalId = setInterval(function() {
+      if (win.window.navigator.onLine) {
+        clearInterval(reloadIntervalId);
+      } else {
+        win.reload();
+      }
+    }, 10 * 1000);
   }
 };
